+++ repos/ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/ktor/server/key/value/KtorReadKeyValueRepoRoutes.kt
@@ -51,7 +51,7 @@ inline fun <reified Key, reified Value> Route.configureReadKeyValueRepoRoutes (
 
     get(keysRoute) {
         val pagination = call.request.queryParameters.extractPagination
-        val reversed = call.getQueryParameterOrSendError(reversedParameterName) ?.toBoolean() ?: false
+        val reversed = call.getQueryParameter(reversedParameterName) ?.toBoolean() ?: false
         val value = call.getQueryParameter(valueParameterName) ?.let {
             valueDeserializer(it)
         }